ai-test-agent/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-system"
version = "0.1.0"
description = "AI Agent System for document processing and generation"
requires-python = ">=3.10,<3.11"
dependencies = [
    "Flask==2.0.1",
    "Werkzeug==2.0.3",
    "Flask-RESTful==0.3.9",
    "Flask-JWT-Extended==4.4.4",
    "Flask-Cors==3.0.10",
    "python-docx==0.8.11",
    "PyPDF2==2.11.1",
    "mistune==2.0.4",
    "transformers==4.28.1",
    "torch==1.13.1",
    "nltk==3.7",
    "spacy==3.5.0",
    "openai==0.27.0",
    "sqlalchemy==1.4.46",
    "psycopg2-binary==2.9.6",
    "python-dotenv==0.20.0",
    "pytest==7.3.1",
    "pytest-cov==4.0.0",
    "gunicorn==20.1.0",
    "docker==6.0.1",
]

[tool.setuptools]
py-modules = []
